Переработка функции сортировки + перенос в utils
