+++ package.json
@@ -12,7 +12,7 @@
   "devDependencies": {
     "css-loader": "5.0.1",
     "eslint": "7.13.0",
-    "eslint-config-htmlacademy": "0.6.0",
+    "eslint-config-htmlacademy": "3.0.0",
     "style-loader": "2.0.0",
     "webpack": "4.44.2",
     "webpack-cli": "3.3.12",
